PHP 8 runtime compatibility fixes (found by running the live site)
Static checks (php -l) only prove the code parses; running the site with
error_reporting=E_ALL + xdebug exposed real runtime breakage. Fixed:

- Dynamic properties (deprecated since 8.2): #[\AllowDynamicProperties]
  added to 25 classes (component models, cmsBlogs, cmsPhoto, cmsUser,
  keyword/rss helpers, bundled IXR library); cmsPage::$site_cfg and
  cmsActions::$inDB declared explicitly. Duplicate attribute in
  cmsUser (added by an earlier commit) removed.
- Fatal TypeError in cms.php: $inConf->timediff is '' by default and
  '' * 3600 is a TypeError on PHP 8 - cast to (int) in 4 places, plus
  rssfeed.
- Fatal TypeError in core/classes/plugin.class.php: array_merge() with
  null config - both arguments cast to (array).
- Deprecated optional-before-required parameter in cmsBlogs::getPosts()
  ($ownerModel now defaults to null).
- Smarty 4.5 deprecation for unregistered functions as modifiers:
  ceil/floor/round/abs/str_repeat/icms_ucfirst registered explicitly.

Verified on the test site: /, /blogs, /forum, /photos, /users,
/catalog, /search, /faq, /admin - HTTP 200, zero PHP errors/warnings.

+++ core/cms.php
@@ -1935,12 +1935,12 @@ class cmsCore {
 	    global $_LANG;
 
 		// формируем входную $date с учетом смещения
-        $date = date('Y-m-d H:i:s', strtotime($date)+($inConf->timediff*3600));
+        $date = date('Y-m-d H:i:s', strtotime($date)+((int)$inConf->timediff*3600));
 
 		// сегодняшняя дата
-		$today     = date('Y-m-d', strtotime(date('Y-m-d H:i:s'))+($inConf->timediff*3600));
+		$today     = date('Y-m-d', strtotime(date('Y-m-d H:i:s'))+((int)$inConf->timediff*3600));
 		// вчерашняя дата
-		$yesterday = date('Y-m-d', strtotime(date('Y-m-d H:i:s'))-(86400)+($inConf->timediff*3600));
+		$yesterday = date('Y-m-d', strtotime(date('Y-m-d H:i:s'))-(86400)+((int)$inConf->timediff*3600));
 
 		// получаем значение даты и времени
 		list($day, $time) = explode(' ', $date);
@@ -1994,7 +1994,7 @@ class cmsCore {
 
 	    global $_LANG;
 
-        $d = date('w', strtotime($date)+(cmsConfig::getConfig('timediff')*3600));
+        $d = date('w', strtotime($date)+((int)cmsConfig::getConfig('timediff')*3600));
 
         $days_week = array($_LANG['SUNDAY'], $_LANG['MONDAY'], $_LANG['TUESDAY'], $_LANG['WEDNESDAY'], $_LANG['THURSDAY'], $_LANG['FRIDAY'], $_LANG['SATURDAY']);
